🚚 refactor: entities/application 신설, 지원 도메인 데이터 추출
[변경 내용]
- entities/application 신설: api/applicationApi(201줄), api/applicationKeys,
  model/apiTypes(304줄), model/types, model/mappers(362줄),
  model/projectPermissionTypes 6개 파일 이동 (~1100줄)
- 잔류 application feature 파일(ui/, hooks/)의 상대 import 및 외부 feature
  (matching, project)의 import 경로를 @/entities/application으로 재작성 (33개 파일)

[변경 이유]
- 지원(application) 데이터/타입/매퍼는 matching(지원 현황), project(지원 모달),
  settings가 공유하는 지원 도메인 데이터라 entities가 올바른 위치
- 이 데이터가 features/application에 있어 matching->application 12건,
  project->application 11건의 cross-feature 참조를 유발하던 최대 결합원
- boundaries 위반 51 -> 30건 (-21): matching->application 12->1,
  project->application 11->1

[고려 사항]
- api/applicationApi가 참조하던 model/projectPermissionTypes도 순수 타입이라
  함께 이동(엔티티가 feature를 역참조하지 않도록)
- UI(ApplicationDetailModal, ApplicationStatsSection 등)와 페이지 훅
  (useApplicationPageData, useProjectsPermissions)은 사용자 시나리오이므로
  features/application에 잔류. 이들은 이동한 데이터를 entity로 import
- 엔티티 내부로 이동한 파일 간 상대참조(applicationApi -> ../model/apiTypes,
  mappers -> ./apiTypes)는 api/model 서브구조 보존으로 그대로 유효
- 남은 features->features 위반: settings->auth(9, auth feature 잔여),
  application->project(7, UI 누수) 등. entities/application 범위 밖
- ProjectApplyModal.ui.test 5건 실패는 기존 이슈로 무관
