+++ src/entities/application/api/applicationKeys.ts
