+++ src/entities/application/model/apiTypes.ts
