+++ src/entities/application/model/projectPermissionTypes.ts
