+++ src/entities/application/model/types.ts
