+++ src/features/application/ui/detail-modal/ModalFormPanel.tsx
@@ -7,8 +7,14 @@ import { OptionButton } from "@/shared/ui/option-button/OptionButton"
 import { OptionButtonGroup } from "@/shared/ui/option-button/OptionButtonGroup"
 import { ProjectTitleCard } from "@/shared/ui/ProjectTitleCard"
 
-import type { ApplicantFormData, FormField } from "../../model/types"
-import type { ApplicantDetail, StatusValue } from "../../model/types"
+import type {
+  ApplicantFormData,
+  FormField,
+} from "@/entities/application/model/types"
+import type {
+  ApplicantDetail,
+  StatusValue,
+} from "@/entities/application/model/types"
 
 const ROLE_LABEL: Record<string, string> = {
   plan: "Plan",
