+++ src/features/matching/model/matchingStatusMapper.ts
@@ -1,9 +1,9 @@
 // 서버 응답 -> 매칭 현황 프론트 타입 변환
 
-import { toRoundNumber } from "@/features/application/model/mappers"
+import { toRoundNumber } from "@/entities/application/model/mappers"
 
+import type { ManagedProjectSummaryResponse } from "@/entities/application/model/apiTypes"
 import type { ProjectMembersResponse } from "@/entities/project/api/matchingProject"
-import type { ManagedProjectSummaryResponse } from "@/features/application/model/apiTypes"
 import type { Part } from "@/shared/model/domain"
 
 import type {
